+++ CHANGELOG.md
@@ -93,6 +93,7 @@
   implemented by user code.
 - The `wb_protected` flag is automatically set for data wrapped in a Ruby
   object when `mark` is not set.
+- `TypedData::class` and `TypedData::class_for` now take a `&Ruby` argument.
 
 ### Deprecated
 - `RString::append` (use `RString::buf_append`).